Remove Jenkins schema generation job trigger
Don't run the schema generation release job when deploying the
consignment API. We don't need to do this every time we update the
API, it's slow, and it will make it harder to reconfigure the API
deployment to allow us to deploy different versions.

For now, we'll start the GraphQL schema Jenkins job manually.