Remove tests of consignments created by an admin user
We don't really have a concept of an admin user yet, so it's better for the
general tests to use a regular user token, to check that realistic frontend
calls would work.

This will also make it easier to make the next change, which is to get the
user's ID from their token instead of a GraphQL parameter, since the admin
user doesn't have a user ID in the tests.