Refactor tests to test for creating multiple files
Considering the only input we need to create a file is the consignment
id, and everything else is auto generated or got from the token, there
seems little point in having a mutation which only creates one file.
I'm going to change this so it takes an optional (default 1) argument
for the number of files you want to create and returns an array with
their IDs. This should help in the front end as well.