Make transferring body a compulsory argument when fetching series
This fixes an authorisation issue, where it was possible to fetch series for all departments by
passing a null body parameter.

The only reason that `body` was an optional parameter was to support future admin functionality,
but that does not exist yet so it's more secure to make the parameter compulsory are revisit this
once we have requirements for the admin functions.