Add error code to GraphQL errors
GraphQL supports error codes to provide more information to the client applications.
We can look for these code Strings in the GraphQL clients to provide better errors
to the user.

For now, the only specialised error is NOT_AUTHORISED, which indicates that the request
tried to do something which the user does not have permission for.