Move authorisation tags to the auth package
Tags are validation steps that can be added to queries and mutations. They are analysed by Sangria
middleware before the query or mutation is run.

We're currently just using them for authorisation: making sure that the user who is calling the
endpoing has permission to do so.

This refactoring will let us distinguish between authorisation validation and any future validation
that is added for other reasons.