Add check to ensure all queries/mutations have an authorisation step
Throw an exception if a query or mutation has been defined without an authorisation
tag to limit access to the query.

This does two things:

- Adds a safety net so that we can't accidentally add a new query with no authorisation,
  and unintentionally make it available to all users
- If we _do_ ever add an endpoint which should be available to all users, we will need
  to add an explicit "authorise all" tag, which makes it easy to find those open endpoints

Note that this just enforces endpoint authorisation. Every endpoint is already protected by
authentication checks, so only users with a valid token can access the API.