Move authorisation Await.result to middleware
This refactoring means that when we add more field authorisation, the risky Await.result is
isolated in a single place. Ideally, we wouldn't have to call Await.result at all, but the
Sangria authorisation middleware doesn't allow us to return a Future.

We'll add a card to the backlog to find a better solution to this. We may have to move the
authorisation code out of the middleware and into the resolvers.