Add functions to get count of total and processed files in consignment
This change adds functions to count both the total number of files in a
consignment and the amount of AVMetadata that has been processed for
those files so far. This is to facilitate the creation of an API
endpoint so we can show TDR users how much processing their uploaded
files have been through so far.