Make antivirus metadata fields non-optional
The antivirus scan generates data for all of these fields, so they shouldn't be
optional. Upgrade to the latest version of the generated database classes, which
makes the fields mandatory in the database, and also make them mandatory in the
GraphQL schema.

The only exception to this is the 'value' field, which is not used and will be
deleted in a later commit.