Pass keycloak deployment to auth utils
Each time we were getting a token, we were creating a keycloak
deployment which contacts the keycloak server to get the oidc
configuration. This is a waste so the auth utils library has been
changed to take an implicit KeycloakConfiguration object. This means
that the config it only got once.