Move client file metadata to file metadata.
Originally, I was going to add a new addFileMetadata field which would
take a list of arguments and then use this to update the client file
metadata but it turns out that this makes the permissions really
complicated because the same query field is used by users, backend
checks and the export. It will be much simpler to keep the current query
structure for client side metadata and change the service so it calls
the file metadata repository.