Clean up test data after every test
Add a new trait which cleans up all test data, and make all of the
tests mix in the trait if they access the test database.

A lot of the individual tests have had to change because they were
depending on shared test data.

This fixes an issue where some tests would fail if only a subset of
the tests were run, because of inter-dependencies. Now you can run all
of the tests, or just one folder, or just one individual test.