Remove unnecessary inheritance from test class
This inheritance wasn't being used for anything, and was making it hard to
debug failing tests, because `GetSeries` is a data class and `TestRequest` is
a ScalaTest helper. This meant that when a test failed, ScalaTest would try
to deserialize `GetSeries`, but it would fail because it couldn't deserialize
some of the private fields on `TestRequest`.