Create foreign keys at same time as table in test database
Merge the `ALTER TABLE` statements with the `CREATE TABLE` statements
rather than keeping them separate. This means the foreign keys are
only added if the tables don't already exist.

This solves a `ConcurrentModificationException` problem that was
causing the tests to fail intermittently. init.sql is run every time
a new database connection is opened, because it's specified as an
init script in the H2 configuration in the tests' application.conf.
If multiple connections were opened in parallel, for example because
Sangria was resolving deferred resolvers in parallel, this sometimes
meant that data was being read by one thread while another thread
ran `ALTER TABLE` on the same table.

This change should mean that init.sql never modifies existing tables,
which should fix the concurrency error.