Make FFID metadata optional and return multiple matches.
There's two things here. The first is to make ffidMetadata optional as if you try to query it when it isn't set, you get an internal server error.

The second is to allow it to return multiple matches. The original code was calling `ffidMetadataAndMatchesRows.toMap` but if you had more than one match, it was trying to create to entries in that map with the same key which you can't have so you ended up with a random ffid match.