Refactor get file metadata method
The getFileMetadata method was getting complicated because all the logic
for getting the FFID and client side metadata was in a single method. If
we were to add the antivirus to this then it would be too messy. I've
refactored this so the client metadata and the FFID metadata comes from
different methods.